+++ src/show.jl
@@ -1,9 +1,9 @@
-export rec2coeff, showpoly
+export rec2coeff, showpoly, showbasis
 
 title_color =   :underline
 name_color  =   :light_blue
 
-function show(io::IO,m::Measure)
+function show(m::Measure)
     printstyled("Measure dλ(t)=w(t)dt\n";color=title_color)
     print("name:\t\t")
     printstyled("$(m.name)\n"; color=name_color)
@@ -20,7 +20,7 @@ function show(io::IO,m::Measure)
     end
 end
 
-function show(io::IO,m::MultiMeasure)
+function show(m::MultiMeasure)
     p = length(m.name)
     printstyled("$p-variate measure dλ(t)= ∏_{i=1}^$p w_i(t)dt_i\n";color=title_color)
     print("name:")
@@ -33,10 +33,10 @@ function show(io::IO,m::MultiMeasure)
         m.symmetric[i] ? col=:green : col=:red
         i==1 ? printstyled("\t$(m.symmetric[i])\n";color=col) : printstyled("\t\t$(m.symmetric[i])\n";color=col)
     end
-    if length(findall(x->typeof(x)!=Dict{Any,Any},m.pars))>=1 print("pars:")
+    if length(findall(x -> typeof(x) != Dict{Any,Any},m.pars)) >= 1 print("pars:")
         for i=1:p
             print("\t\t")
-            if typeof(m.pars[i])!=Dict{Any,Any}
+            if typeof(m.pars[i]) != Dict{Any,Any}
                 for (i,d) in enumerate(m.pars[i])
                     i==1 ? print("$d") : print(", $d")
                 end
@@ -48,7 +48,7 @@ function show(io::IO,m::MultiMeasure)
     end
 end
 
-function show(io::IO,op::OrthoPoly;showmeasure::Bool=true)
+function show(op::OrthoPoly; showmeasure::Bool=true)
     printstyled("\nUnivariate orthogonal polynomials\n";color=title_color)
     print("name:"); printstyled("\t\t$(op.name)\n";color=name_color)
     print("degree:"); printstyled("\t\t$(op.deg)\n")
@@ -60,11 +60,11 @@ function show(io::IO,op::OrthoPoly;showmeasure::Bool=true)
     n<length(op.α) ? print("...\n") : print("\n")
     if showmeasure
         print("\n")
-        show(io,op.meas)
+        show(op.meas)
     end
 end
 
-function show(io::IO,q::Quad;showmeasure::Bool=true)
+function show(q::Quad; showmeasure::Bool=true)
     printstyled("\nQuadrature rule\n";color=title_color)
     print("name:"); printstyled("\t\t$(q.name)\n";color=name_color)
     print("N:"); printstyled("\t\t$(q.Nquad)\n")
@@ -75,37 +75,35 @@ function show(io::IO,q::Quad;showmeasure::Bool=true)
     n<q.Nquad ? print("...\n") : print("\n")
     if showmeasure
         print("\n")
-        show(io,q.meas)
+        show(q.meas)
     end
 end
 
-function show(io::IO,opq::OrthoPolyQ;showmeasure::Bool=false)
-    show(io,opq.op)
-    show(io,opq.quad;showmeasure=showmeasure)
+function show(opq::OrthoPolyQ; showmeasure::Bool=false)
+    show(opq.op)
+    show(opq.quad; showmeasure=showmeasure)
 end
 
-function show(io::IO,mop::MultiOrthoPoly;showmeasure::Bool=false)
+function show(mop::MultiOrthoPoly; showmeasure::Bool=false)
     p = length(mop.name)
-    printstyled("\n$p-variate orthogonal polynomials\n";color=title_color)
+    printstyled("\n$p-variate orthogonal polynomials\n"; color = title_color)
     print("name:");
-    [ printstyled("\t\t$(mop.name[i])\n";color=name_color) for i=1:p ]
+    [ printstyled("\t\t$(mop.name[i])\n"; color = name_color) for i = 1:p ]
     print("deg:"); printstyled("\t\t$(mop.deg)\n")
     print("dim:"); printstyled("\t\t$(mop.dim)\n")
     print("ind:");
-    mop.dim<=7 ? n=mop.dim : n=7
-    # printstyled("\t$(mop.ind[1,:])\n")
-    [ printstyled("\t\t$(mop.ind[i,:])\n") for i=1:n ]
-    n<mop.dim ? print("\t\t...\n") : print("\n")
-
+    mop.dim <= 7 ? n = mop.dim : n = 7
+    [ printstyled("\t\t$(mop.ind[i,:])\n") for i = 1:n ]
+    n < mop.dim ? print("\t\t...\n") : print("\n")
     print("\n")
-    showmeasure ? show(io,mop.meas) : ()
+    showmeasure && show(mop.meas)
 end
 
-function show(io::IO,t::Tensor)
-    printstyled("\n$(t.dim)-dimensional tensor\n";color=title_color)
+function show(t::Tensor)
+    printstyled("\n$(t.dim)-dimensional tensor\n"; color = title_color)
     print("dim:"); printstyled("\t\t$(t.dim)\n")
     print("nonzeros:"); printstyled("\t$(length(t.T.nzind))\n")
-    show(io,t.op;showmeasure=false)
+    show(t.op; showmeasure = false)
 end
 
 """
@@ -145,60 +143,62 @@ rec2coeff(α,β) = rec2coeff(length(α),α,β)
 """
 ```
 showpoly(coeffs::Vector{Float64};sym::String,digits::Integer)
-showpoly(coeffs::Vector{Vector{Float64}};sym::String,digits::Integer)
 ```
 Show the monic polynomial with coefficients `coeffs` in a human readable way.
 They keyword `sym` sets the name of the variable, and `digits` controls the number of shown digits.
-```@repl
+```jldoctest
+julia> using PolyChaos
+
 julia> showpoly([1.2, 2.3, 3.4456])
 x^3 + 3.45x^2 + 2.3x + 1.2
 julia> showpoly([1.2, 2.3, 3.4456], sym="t", digits=2)
 t^3 + 3.45t^2 + 2.3t + 1.2
 ```
 
+```
+showpoly(d::Integer,α::Vector{Float64},β::Vector{Float64}; sym::String,digits::Integer)
+showpoly(d::Range,α::Vector{Float64},β::Vector{Float64};sym::String,digits::Integer) where Range <: OrdinalRange
+```
+Show the monic polynomial of degree/range `d` that has the recurrence coefficients `α`, `β`.
+```jldoctest
+julia> using PolyChaos
+
+julia> α, β = rm_hermite(10)
+([0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], [1.77245, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5])
+julia> showpoly(3,α,β)
+x^3 - 1.5x
+
+julia> showpoly(0:2:10,α,β)
+1
+x^2 - 0.5
+x^4 - 3.0x^2 + 0.75
+x^6 - 7.5x^4 + 11.25x^2 - 1.88
+x^8 - 14.0x^6 + 52.5x^4 - 52.5x^2 + 6.56
+x^10 - 22.5x^8 + 157.5x^6 - 393.75x^4 + 295.31x^2 - 29.53
+```
+
 Tailored to types from `PolyChaos.jl`
 ```
-showpoly(d::Integer,α::Vector{Float64},β::Vector{Float64};upto::Bool,sym::String,digits::Integer)
-showpoly(d::Integer,op::OrthoPoly;upto::Bool=true,sym::String,digits::Integer)
-showpoly(d::Integer,opq::OrthoPolyQ;upto::Bool=true,sym::String,digits::Integer)
+showpoly(d::Union{Integer,Range},op::OrthoPoly;sym::String,digits::Integer) where Range <: OrdinalRange
+showpoly(d::Union{Integer,Range},opq::OrthoPolyQ;sym::String,digits::Integer) where Range <: OrdinalRange
 ```
-Show the monic orthogonal polynomials with recurrence coefficients `(α,β)` up to degree `d`.
-Setting the keyword `upto` to false prints the monic polynomial of degree equal to `d`.
-```@repl
+Show the monic polynomial of degree/range `d` of an `OrthoPoly`/`OrthoPolyQ`.
+```jldoctest
+julia> using PolyChaos
+
 julia> op = OrthoPoly("gaussian",5);
+
 julia> showpoly(3,op)
-1
-x
-x^2 - 1.0
 x^3 - 3.0x
 
-julia> showpoly(3,op; upto=false)
-x^3 - 3.0x
-
-julia> showpoly(3,op; sym="t")
+julia> showpoly(0:op.deg,op; sym="t")
 1
 t
 t^2 - 1.0
 t^3 - 3.0t
+t^4 - 6.0t^2 + 3.0
+t^5 - 10.0t^3 + 15.0t
 ```
-
-The following function calls show all orthogonal polynomials given `(α,β)`.
-```
-showpoly(α::Vector{Float64},β::Vector{Float64};sym::String,digits::Integer)
-showpoly(op::Union{OrthoPoly,OrthoPolyQ};sym::String,digits::Integer)
-```
-
-```@repl
-julia> showpoly(op; sym="y")
-1
-y
-y^2 - 1.0
-y^3 - 3.0y
-y^4 - 6.0y^2 + 3.0
-y^5 - 10.0y^3 + 15.0y
-y^6 - 15.0y^4 + 45.0y^2 - 15.0
-```
-
 [Thanks @pfitzseb for providing this functionality.](https://discourse.julialang.org/t/how-to-define-verbose-output-for-a-polynomial/21317/5)
 
 """
@@ -216,21 +216,62 @@ function showpoly(coeffs::Vector{Float64};sym::String="x",digits::Integer=2)
     print(io, '\n')
 end
 
-function showpoly(coeffs::Vector{Vector{Float64}};sym::String="x",digits::Integer=2)
-    print("1\n")
-    map(c->showpoly(c,sym=sym,digits=digits),coeffs)
+function showpoly(d::Integer,α::Vector{Float64},β::Vector{Float64}; sym::String="x",digits::Integer=2)
+    @assert d >= 0 "degree has to be non-negative."
+    d == 0 && return print("1\n")
+    showpoly(rec2coeff(d,α,β)[end],sym=sym,digits=digits)
+end
+
+function showpoly(d::Range,α::Vector{Float64},β::Vector{Float64};sym::String="x",digits::Integer=2) where Range <: OrdinalRange
+    map(c->showpoly(c,α,β;sym=sym,digits=digits),d)
     print()
 end
 
-function showpoly(d::Integer,α::Vector{Float64},β::Vector{Float64};upto::Bool=true,sym::String="x",digits::Integer=2)
-    @assert d >= 0 "degree has to be non-negative."
-    d == 0 && return print("1")
-    cfs = upto ? rec2coeff(d,α,β) : rec2coeff(d,α,β)[end]
-    showpoly(cfs,sym=sym,digits=digits)
+showpoly(d::Union{Integer,Range},op::OrthoPoly;sym::String="x",digits::Integer=2) where Range <: OrdinalRange = showpoly(d,op.α,op.β;sym=sym,digits=digits)
+showpoly(d::Union{Integer,Range},opq::OrthoPolyQ;sym::String="x",digits::Integer=2) where Range <: OrdinalRange = showpoly(d,opq.op;sym=sym,digits=digits)
+
+
+"""
+```
+showbasis(α::Vector{Float64},β::Vector{Float64};sym::String,digits::Integer)
+```
+Show all basis polynomials given the recurrence coefficients `α`, `β`.
+They keyword `sym` sets the name of the variable, and `digits` controls the number of shown digits.
+```jldoctest
+julia> using PolyChaos
+
+julia> α, β = rm_hermite(5);
+
+julia> showbasis(α,β)
+1
+x
+x^2 - 0.5
+x^3 - 1.5x
+x^4 - 3.0x^2 + 0.75
+x^5 - 5.0x^3 + 3.75x
+```
+
+Tailored to types from `PolyChaos.jl`
+```
+showbasis(op::OrthoPoly;sym::String,digits::Integer)
+showbasis(opq::OrthoPolyQ;sym::String,digits::Integer)
+```
+Show all basis polynomials of an `OrthoPoly`/`OrthoPolyQ`.
+```jldoctest
+julia> using PolyChaos
+
+julia> op = OrthoPoly("legendre",4);
+
+julia> showbasis(op)
+1
+x
+x^2 - 0.33
+x^3 - 0.6x
+x^4 - 0.86x^2 + 0.09
+```
+"""
+function showbasis(α::Vector{Float64},β::Vector{Float64};sym::String="x",digits::Integer=2)
+    showpoly(0:length(α),α,β;sym=sym,digits=digits)
 end
-
-showpoly(d::Integer,op::OrthoPoly;upto::Bool=true,sym::String="x",digits::Integer=2) = showpoly(d,op.α,op.β;upto=upto,sym=sym,digits=digits)
-showpoly(d::Integer,opq::OrthoPolyQ;upto::Bool=true,sym::String="x",digits::Integer=2) = showpoly(d,opq.op;upto=upto,sym=sym,digits=digits)
-
-showpoly(α::Vector{Float64},β::Vector{Float64};sym::String="x",digits::Integer=2) = showpoly(length(α),α,β;sym=sym,digits=digits)
-showpoly(op::Union{OrthoPoly,OrthoPolyQ};sym::String="x",digits::Integer=2) = showpoly(size(coeffs(op),1),op;sym=sym,digits=digits)
+showbasis(op::OrthoPoly;sym::String="x",digits::Integer=2) = showbasis(op.α[1:op.deg],op.β[1:op.deg];sym=sym,digits=digits)
+showbasis(opq::OrthoPolyQ;sym::String="x",digits::Integer=2) = showbasis(opq.op;sym=sym,digits=digits)